separate server features into a plugin
