+++ plugins/typed-route/type.ts
@@ -1,1 +1,1 @@
-export type RoutesType = `/dynamic` | `/link` | `/new` | `/other` | `/other/static` | `/serversideprops` | `/serversideprops/static` | `/serversideprops/static/undefined` | `/serversideprops/undefined` | `/showoff` | `/style-page` | `/svg` | `/` | `/dynamic/static/${string}/` | `/dynamic/static/[module]` | `/static/${string}/` | `/static/[id]` | `/test/foo/${string}/` | `/test/foo/[bar]` | `/test/${string}/` | `/test/[id]` | `/${string}/other` | `/[segment]/other`
+export type RoutesType = `/dynamic` | `/link` | `/other` | `/other/static` | `/serversideprops` | `/serversideprops/static` | `/serversideprops/static/undefined` | `/serversideprops/undefined` | `/showoff` | `/style-page` | `/svg` | `/` | `/dynamic/static/${string}/` | `/dynamic/static/[module]` | `/static/${string}/` | `/static/[id]` | `/test/foo/${string}/` | `/test/foo/[bar]` | `/test/${string}/` | `/test/[id]` | `/${string}/other` | `/[segment]/other`
