+++ src/pages/index.tsx
@@ -2,8 +2,7 @@ import { TestElement } from "./test";
 import { useSession, GetSession } from "bunext-js/session";
 import { generateRandomString } from "../../features/utils";
 import { TestServerElement2 } from "./serverElement";
-import "@static/index.css";
-import "@static/style.css";
+
 import { Head } from "bunext-js/head";
 import { useEffect } from "react";
