--- tailwind.config.ts
@@ -1,10 +1,0 @@
-import type { Config } from "tailwindcss";
-
-export default {
- content: ["./src/**/*.tsx"],
- theme: {
-  extend: {},
- },
- important: true,
- plugins: [],
-} satisfies Config;
